requirements.txt
ultralytics
streamlit
opencv-python-headless
pillow
plotly
fpdf
